# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Return Missionary Association Api",
    "description": "This application stores resources for the return missionary application",
    "version": "1.0.0",
    "contact": {
      "email": "easyboy@gmail.com"
    }
  },
  "host": "rm-association.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
...
  "paths": {
...
    "/attendance/{attendanceid}": {
      "get": {
        "tags": ["attendance"],
        "summary": "Get attendance by Id",
        "description": "Returns a single member attendance",
        "operationalId": "getAttendanceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of attendance to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["attendance"],
        "summary": "Deletes an attendance",
        "description": "delete attendance",
        "operationId": "deleteAttendance",
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "Attendance id to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": ""
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid  attendance Id"
          }
        }
      },
      "put": {
        "tags": ["attendance"],
        "summary": "Update an attendance",
        "description": "Attendance to update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "objectId"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "eventId": {
                  "an obectId of event": "284729279",
                  "required": true
                },
                "memberId": {
                  "an objectid of member": "jsdjs372377",
                  "required": true
                },
                "status": {
                  "present": "present",
                  "required": true
                },
                "timestamp": {
                  "date": ""
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/events/{id}": {
      "get": {
        "tags": ["event"],
        "summary": "Get event by Id",
        "description": "Returns a single member attendance",
        "operationalId": "getEventById",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "description": "ID of event to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["attendance"],
        "summary": "Deletes an attendance",
        "description": "delete attendance",
        "operationId": "deleteAttendance",
        "parameters": [
          {
            "name": "attendanceId",
            "in": "path",
            "description": "Attendance id to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": ""
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid  attendance Id"
          }
        }
      },
      "put": {
        "tags": ["attendance"],
        "summary": "Update an attendance",
        "description": "Attendance to update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "objectId"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "name": "nameofevent",
                  "required": true
                },
                "location": {
                  "location": "locationofevent",
                  "required": true
                },
                "description": {
                  "description": "describe what the event is all about",
                  "required": true
                },
                "date": {
                  "date": "date format"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}